Move FeatureFlags enum to common, simplify isFeatureFlagEnabled implementation
