--- packages/backend/src/services/FeatureFlagService/FeatureFlagService.mock.ts
@@ -1,6 +1,0 @@
-import { SessionUser } from '@lightdash/common';
-
-export const user: Pick<SessionUser, 'userUuid' | 'organizationUuid'> = {
-    userUuid: 'userUuid',
-    organizationUuid: 'organizationUuid',
-};
